requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
cachetools==5.5.2
captcha==0.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
easyocr==1.7.2
filelock==3.17.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jiter==0.8.2
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.3
numpy==2.2.3
openai==1.64.0
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pillow==11.1.0
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.8.0
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.25.2
scipy==1.15.2
selenium==4.29.0
setuptools==75.8.2
shapely==2.0.7
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sympy==1.13.1
tifffile==2025.2.18
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
zhipuai==2.1.5.20250106
